Making baselineDef more objects like
1. split definition and declaration to .h and .cc file
2. Move common input labels as private members
3. Ideally, each analysis jsut need to define their own baseline class
and modify it locally, instead of all messing up this code
4. add nTop objects to the baseline, so that we can get rid of the global
type3ptr, which is dangerous
